+++ ReleaseNotes/release_notes.md
@@ -1,4 +1,5 @@
 ## New features: 
+  1. some massage for `BernsteinEven` and `PositiveEven`
 
 ## Backward incompatible: